+++ src/main/java/com/fzi/externalcontrol/impl/ExternalControlInstallationNodeContribution.java
@@ -25,6 +25,9 @@
 
 package com.fzi.externalcontrol.impl;
 
+import static com.fzi.externalcontrol.impl.ExternalControlInstallationNodeContribution.DEFAULT_PORT;
+import static com.fzi.externalcontrol.impl.ExternalControlInstallationNodeContribution.PORT_NR;
+
 import com.ur.urcap.api.contribution.InstallationNodeContribution;
 import com.ur.urcap.api.contribution.installation.InstallationAPIProvider;
 import com.ur.urcap.api.domain.data.DataModel;
@@ -34,23 +37,16 @@ import com.ur.urcap.api.domain.userinteraction.keyboard.KeyboardInputFactory;
 import com.ur.urcap.api.domain.userinteraction.keyboard.KeyboardTextInput;
 
 public class ExternalControlInstallationNodeContribution implements InstallationNodeContribution {
-  private static final String HOST_IP = "host_ip";
-  private static final String PORT_NR = "port_nr";
-  private static final String NAME = "name";
-  private String urScriptProgram = "";
-  private static final String DEFAULT_IP = "192.168.56.1";
-  private static final String DEFAULT_PORT = "50002";
-  private static final String DEFAULT_NAME = DEFAULT_IP;
-  private DataModel model;
-  private final ExternalControlInstallationNodeView view;
-  private final KeyboardInputFactory keyboardFactory;
-  private int instanceCounter = 0;
-
-  private PitascCaller pitascCaller;
-  private static final String PITASC_PORT_NR = "pitasc_port";
-  private static final String PITASC_DEFAULT_PORT = "40404";
-  private static final String PITASC_FILE = "pitasc_file";
-  private static final String PITASC_DEFAULT_FILE = "default.xml";
+  protected static final String HOST_IP = "host_ip";
+  protected static final String PORT_NR = "port_nr";
+  protected static final String NAME = "name";
+  protected String urScriptProgram = "";
+  protected static final String DEFAULT_IP = "192.168.56.1";
+  protected static final String DEFAULT_PORT = "50002";
+  protected static final String DEFAULT_NAME = DEFAULT_IP;
+  protected DataModel model;
+  protected final ExternalControlInstallationNodeView view;
+  protected final KeyboardInputFactory keyboardFactory;
 
  
   public ExternalControlInstallationNodeContribution(InstallationAPIProvider apiProvider,
@@ -75,9 +71,6 @@ public class ExternalControlInstallationNodeContribution implements Installation
   public void generateScript(ScriptWriter writer) {
     RequestProgram sender = new RequestProgram(getHostIP(), getCustomPort());
     urScriptProgram = sender.sendCommand("request_program\n");
-
-    pitascCaller = new PitascCaller(getHostIP(), getPitascPort());
-    pitascCaller.appendInstallationLines(writer, getPitascFile());
   }
 
   // IP helper functions
@@ -113,147 +106,14 @@ public class ExternalControlInstallationNodeContribution implements Installation
     };
   }
 
-  // port helper functions
-  public void setHostPort(String port) {
-    if ("".equals(port)) {
-      resetToDefaultPort();
-    } else {
-      model.set(PORT_NR, port);
-    }
-  }
 
-  public String getCustomPort() {
-    return model.get(PORT_NR, DEFAULT_PORT);
-  }
-
-  private void resetToDefaultPort() {
-    model.set(PORT_NR, DEFAULT_PORT);
-  }
-
-  public KeyboardTextInput getInputForPortTextField() {
-    KeyboardTextInput keyboInput = keyboardFactory.createStringKeyboardInput();
-    keyboInput.setInitialValue(getCustomPort());
-    return keyboInput;
-  }
-
-  public KeyboardInputCallback<String> getCallbackForPortTextField() {
-    return new KeyboardInputCallback<String>() {
-      @Override
-      public void onOk(String value) {
-        setHostPort(value);
-        view.UpdatePortTextField(value);
-      }
-    };
-  }
-
-  // name helper functions
-  public void setName(String name) {
-    if ("".equals(name)) {
-      resetToDefaultName();
-    } else {
-      model.set(NAME, name);
-    }
-  }
-
-  public String getName() {
-    return model.get(NAME, DEFAULT_NAME);
-  }
-
-  private void resetToDefaultName() {
-    model.set(NAME, DEFAULT_NAME);
-  }
-
-  public KeyboardTextInput getInputForNameTextField() {
-    KeyboardTextInput keyboInput = keyboardFactory.createStringKeyboardInput();
-    keyboInput.setInitialValue(getName());
-    return keyboInput;
-  }
-
-  public KeyboardInputCallback<String> getCallbackForNameTextField() {
-    return new KeyboardInputCallback<String>() {
-      @Override
-      public void onOk(String value) {
-        setName(value);
-        view.UpdateNameTextField(value);
-      }
-    };
-  }
-
-  // port helper functions
-  public void setPitascPort(String port) {
-    if ("".equals(port)) {
-      resetToDefaultPitascPort();
-    } else {
-      model.set(PITASC_PORT_NR, port);
-    }
-  }
-
-  public String getPitascPort() {
-    return model.get(PITASC_PORT_NR, PITASC_DEFAULT_PORT);
-  }
-
-  private void resetToDefaultPitascPort() {
-    model.set(PITASC_PORT_NR, PITASC_DEFAULT_PORT);
-  }
-  
-  public KeyboardTextInput getInputForPitascPortTextField() {
-    KeyboardTextInput keyboInput = keyboardFactory.createStringKeyboardInput();
-    keyboInput.setInitialValue(getPitascPort());
-    return keyboInput;
-  }
-
-  public KeyboardInputCallback<String> getCallbackForPitascPortTextField() {
-    return new KeyboardInputCallback<String>() {
-      @Override
-      public void onOk(String value) {
-        setPitascPort(value);
-        view.UpdatePitascPortTextField(value);
-      }
-    };
-  }
-
-  // port helper functions
-  public void setPitascFile(String file) {
-    if ("".equals(file)) {
-      resetToDefaultPitascFile();
-    } else {
-      model.set(PITASC_FILE, file);
-    }
-  }
-
-  public String getPitascFile() {
-    return model.get(PITASC_FILE, PITASC_DEFAULT_FILE);
-  }
-
-  private void resetToDefaultPitascFile() {
-    model.set(PITASC_FILE, PITASC_DEFAULT_FILE);
-  }
-  
-  public KeyboardTextInput getInputForPitascFileTextField() {
-    KeyboardTextInput keyboInput = keyboardFactory.createStringKeyboardInput();
-    keyboInput.setInitialValue(getPitascFile());
-    return keyboInput;
-  }
-
-  public KeyboardInputCallback<String> getCallbackForPitascFileTextField() {
-    return new KeyboardInputCallback<String>() {
-      @Override
-      public void onOk(String value) {
-        setPitascFile(value);
-        view.UpdatePitascFileTextField(value);
-      }
-    };
-  }
-  
   public String getUrScriptProgram() {
     return urScriptProgram;
   }
   
-  public int IncrementInstanceCounter() {
-    return this.instanceCounter++;
+  public String getCustomPort() {
+	return model.get(PORT_NR, DEFAULT_PORT);
   }
 
-  public PitascCaller getPitascCaller() {
-	  return pitascCaller;
-  }
+  
 }
