+++ src/main/java/com/fzi/externalcontrol/impl/ExternalControlInstallationNodeView.java
@@ -32,7 +32,6 @@ import java.awt.Component;
 import java.awt.Dimension;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
-
 import javax.swing.Box;
 import javax.swing.JLabel;
 import javax.swing.JPanel;
