+++ src/main/java/com/fzi/externalcontrol/impl/IPAControlInstallationNodeService.java
@@ -1,0 +1,12 @@
+package com.fzi.externalcontrol.impl;
+
+import java.util.Locale;
+
+public class IPAControlInstallationNodeService extends ExternalControlInstallationNodeService {
+	
+	  @Override
+	  public String getTitle(Locale locale) {
+		    return "External Pitasc";
+		  }
+
+}
