+++ src/main/java/com/fzi/externalcontrol/impl/IPAControlInstallationNodeView.java
@@ -1,0 +1,5 @@
+package com.fzi.externalcontrol.impl;
+
+public class IPAControlInstallationNodeView extends ExternalControlInstallationNodeView {
+
+}
