+++ src/main/java/com/fzi/externalcontrol/impl/IPAControlProgramNodeView.java
@@ -1,0 +1,5 @@
+package com.fzi.externalcontrol.impl;
+
+public class IPAControlProgramNodeView {
+
+}
